# src/main/resources/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Financial System API
  version: 1.0.0
  description: API for managing users and their accounts
tags:
  - name: Auth api
  - name: User api
  - name: Account api
paths:
  /auth/login:
    post:
      summary: Authenticate and get JWT token
      operationId: login
      tags:
        - Auth api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'

  /users:
    get:
      summary: Search users by params
      operationId: findUsersBy
      tags:
        - User api
      parameters:
        - in: query
          name: pageable
          required: false
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pageable'
        - in: query
          name: userSearchParams
          required: false
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSearchParams'
      responses:
        200:
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSearchResponse'

  /users/{id}:
    get:
      summary: Get user profile by id
      operationId: getUserById
      tags:
        - User api
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserByIdResponse'
  
  /users/{userId}/emails:
    put:
      summary: Create or update user emails
      operationId: createOrUpdateUserEmails
      tags:
        - User api
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEmailsDto'
      responses: 
        200:
          description: User emails
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEmailsDto'

  /users/{userId}/phones:
    put:
      summary: Create or update user phones
      operationId: createOrUpdateUserPhones
      tags:
        - User api
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPhonesDto'
      responses:
        200:
          description: User phones
          content:
            application/json:
              schema:
               $ref: '#/components/schemas/UserPhonesDto'

  /accounts/transfer:
    post:
      summary: Transfer money
      operationId: transfer
      tags:
        - Account api
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        200:
          description: Transfer completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'

components:
  schemas:
    Pageable:
      type: object
      properties:
        page:
          description: The page number (0-based)
          type: integer
          default: 0
        size:
          description: The number of items per page
          type: integer
          default: 10

    LoginRequest:
      type: object
      required:
        - password
      properties:
        phone:
          type: string
          description: Phone number
        email:
          type: string
          description: Email
        password:
          type: string
          format: password
      oneOf:
        - required:
            - phone
        - required:
            - email

    LoginResponse:
      type: object
      properties:
        token:
          type: string

    UserSearchParams:
      type: object
      properties:
        dateOfBirth:
          type: string
          format: date
        phone:
          type: string
        name:
          type: string
        email:
          type: string

    UserSearchResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'

    UserByIdResponse:
      type: object
      properties:
        content:
          $ref: '#/components/schemas/UserDto'

    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        dateOfBirth:
          type: string
          pattern: '^(0[1-9]|[12][0-9]|3[01])\.(0[1-9]|1[0-2])\.\d{4}$'
        emails:
          type: array
          items:
            type: string
        phones:
          type: array
          items:
            type: string
        account:
          $ref: '#/components/schemas/AccountDTO'

    UserEmailsDto:
      type: object
      required:
        - emails
      properties:
        emails:
          type: array
          minItems: 1
          items:
            type: string

    UserPhonesDto:
      type: object
      required:
        - phones
      properties:
        phones:
          type: array
          minItems: 1
          items:
            type: string
            pattern: '^\d{11}$'

    TransferRequest:
      type: object
      properties:
        toUserId:
          type: integer
          format: int64
        value:
          type: number
          format: number

    AccountDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        balance:
          type: number
          format: number

    ErrorResponse:
      type: object
      properties:
        message:
          type: string

    TransferResponse:
      type: object
      properties:
        message:
          type: string